split context file
