+++ modules/core/src/main/scala/sangria/schema/Context.scala
@@ -1,218 +1,14 @@
 package sangria.schema
 
-import language.{higherKinds, implicitConversions}
-import sangria.execution._
-import sangria.marshalling._
 import sangria.ast.SourceMapper
-import sangria.{ast, introspection}
+import sangria.execution._
 import sangria.execution.deferred.Deferred
-import sangria.streaming.SubscriptionStream
+import sangria.marshalling._
 import sangria.util.Cache
+import sangria.{ast, introspection}
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.language.implicitConversions
 import scala.reflect.ClassTag
-import scala.util.{Failure, Try}
-import scala.util.control.NonFatal
-
-sealed trait Action[+Ctx, +Val] {
-  def map[NewVal](fn: Val => NewVal)(implicit ec: ExecutionContext): Action[Ctx, NewVal]
-}
-sealed trait LeafAction[+Ctx, +Val] extends Action[Ctx, Val] {
-  def map[NewVal](fn: Val => NewVal)(implicit ec: ExecutionContext): LeafAction[Ctx, NewVal]
-}
-sealed trait ReduceAction[+Ctx, +Val] extends Action[Ctx, Val] {
-  def map[NewVal](fn: Val => NewVal)(implicit ec: ExecutionContext): LeafAction[Ctx, NewVal]
-}
-
-object ReduceAction {
-  implicit def futureAction[Ctx, Val](value: Future[Val]): ReduceAction[Ctx, Val] = FutureValue(
-    value)
-  implicit def tryAction[Ctx, Val](value: Try[Val]): ReduceAction[Ctx, Val] = TryValue(value)
-  implicit def defaultAction[Ctx, Val](value: Val): ReduceAction[Ctx, Val] = Value(value)
-}
-
-object Action extends LowPrioActions {
-  def sequence[Ctx, Val](actions: Seq[LeafAction[Ctx, Val]]): SequenceLeafAction[Ctx, Val] =
-    SequenceLeafAction[Ctx, Val](actions)
-
-  def apply[Ctx, Val](a: Action[Ctx, Val]): Action[Ctx, Val] = a
-
-  implicit def deferredAction[Ctx, Val](value: Deferred[Val]): LeafAction[Ctx, Val] = DeferredValue(
-    value)
-  implicit def tryAction[Ctx, Val](value: Try[Val]): LeafAction[Ctx, Val] = TryValue(value)
-}
-
-trait LowPrioActions extends LowestPrioActions {
-  implicit def deferredFutureAction[Ctx, Val](value: Future[Deferred[Val]]): LeafAction[Ctx, Val] =
-    DeferredFutureValue(value)
-}
-
-trait LowestPrioActions {
-  implicit def futureAction[Ctx, Val](value: Future[Val]): LeafAction[Ctx, Val] = FutureValue(value)
-  implicit def defaultAction[Ctx, Val](value: Val): LeafAction[Ctx, Val] = Value(value)
-}
-
-object LeafAction {
-  def sequence[Ctx, Val](actions: Seq[LeafAction[Ctx, Val]]): SequenceLeafAction[Ctx, Val] =
-    SequenceLeafAction[Ctx, Val](actions)
-
-  def apply[Ctx, Val](a: LeafAction[Ctx, Val]): LeafAction[Ctx, Val] = a
-}
-
-case class Value[Ctx, Val](value: Val) extends LeafAction[Ctx, Val] with ReduceAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): LeafAction[Ctx, NewVal] =
-    try Value(fn(value))
-    catch {
-      case NonFatal(e) => TryValue(Failure(e))
-    }
-}
-
-case class TryValue[Ctx, Val](value: Try[Val])
-    extends LeafAction[Ctx, Val]
-    with ReduceAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): TryValue[Ctx, NewVal] =
-    TryValue(value.map(fn))
-}
-
-case class PartialValue[Ctx, Val](value: Val, errors: Vector[Throwable])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): LeafAction[Ctx, NewVal] =
-    try PartialValue(fn(value), errors)
-    catch {
-      case NonFatal(e) => TryValue(Failure(e))
-    }
-}
-
-case class FutureValue[Ctx, Val](value: Future[Val])
-    extends LeafAction[Ctx, Val]
-    with ReduceAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): FutureValue[Ctx, NewVal] =
-    FutureValue(value.map(fn))
-}
-
-case class PartialFutureValue[Ctx, Val](value: Future[PartialValue[Ctx, Val]])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): PartialFutureValue[Ctx, NewVal] =
-    PartialFutureValue(value.map(_.map(fn) match {
-      case v: PartialValue[Ctx, NewVal] => v
-      case TryValue(Failure(e)) => throw e
-      case v => throw new IllegalStateException("Unexpected result from `PartialValue.map`: " + v)
-    }))
-}
-
-case class DeferredValue[Ctx, Val](value: Deferred[Val]) extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): DeferredValue[Ctx, NewVal] =
-    DeferredValue(MappingDeferred(value, (v: Val) => (fn(v), Vector.empty)))
-
-  def mapWithErrors[NewVal](fn: Val => (NewVal, Vector[Throwable])): DeferredValue[Ctx, NewVal] =
-    DeferredValue(MappingDeferred(value, fn))
-}
-
-case class DeferredFutureValue[Ctx, Val](value: Future[Deferred[Val]])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): DeferredFutureValue[Ctx, NewVal] =
-    DeferredFutureValue(value.map(MappingDeferred(_, (v: Val) => (fn(v), Vector.empty))))
-
-  def mapWithErrors[NewVal](fn: Val => (NewVal, Vector[Throwable]))(implicit
-      ec: ExecutionContext): DeferredFutureValue[Ctx, NewVal] =
-    DeferredFutureValue(value.map(MappingDeferred(_, fn)))
-}
-
-case class SequenceLeafAction[Ctx, Val](value: Seq[LeafAction[Ctx, Val]])
-    extends LeafAction[Ctx, Seq[Val]] {
-  override def map[NewVal](fn: Seq[Val] => NewVal)(implicit
-      ec: ExecutionContext): MappedSequenceLeafAction[Ctx, Val, NewVal] =
-    new MappedSequenceLeafAction[Ctx, Val, NewVal](this, fn)
-}
-
-class MappedSequenceLeafAction[Ctx, Val, NewVal](
-    val action: SequenceLeafAction[Ctx, Val],
-    val mapFn: Seq[Val] => NewVal)
-    extends LeafAction[Ctx, NewVal] {
-  override def map[NewNewVal](fn: NewVal => NewNewVal)(implicit
-      ec: ExecutionContext): MappedSequenceLeafAction[Ctx, Val, NewNewVal] =
-    new MappedSequenceLeafAction[Ctx, Val, NewNewVal](action, v => fn(mapFn(v)))
-}
-
-class UpdateCtx[Ctx, Val](val action: LeafAction[Ctx, Val], val nextCtx: Val => Ctx)
-    extends Action[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): MappedUpdateCtx[Ctx, Val, NewVal] =
-    new MappedUpdateCtx[Ctx, Val, NewVal](action, nextCtx, fn)
-}
-
-class MappedUpdateCtx[Ctx, Val, NewVal](
-    val action: LeafAction[Ctx, Val],
-    val nextCtx: Val => Ctx,
-    val mapFn: Val => NewVal)
-    extends Action[Ctx, NewVal] {
-  override def map[NewNewVal](fn: NewVal => NewNewVal)(implicit
-      ec: ExecutionContext): MappedUpdateCtx[Ctx, Val, NewNewVal] =
-    new MappedUpdateCtx[Ctx, Val, NewNewVal](action, nextCtx, v => fn(mapFn(v)))
-}
-
-object UpdateCtx {
-  def apply[Ctx, Val](action: LeafAction[Ctx, Val])(newCtx: Val => Ctx): UpdateCtx[Ctx, Val] =
-    new UpdateCtx(action, newCtx)
-}
-
-private[sangria] case class SubscriptionValue[Ctx, Val, S[_]](
-    source: S[Any],
-    stream: SubscriptionStream[S])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): SubscriptionValue[Ctx, NewVal, S] =
-    throw new IllegalStateException(
-      "`map` is not supported subscription actions. Action is only intended for internal use.")
-}
-
-case class ProjectionName(name: String) extends FieldTag
-case object ProjectionExclude extends FieldTag
-
-trait Projector[Ctx, Val, Res] extends (Context[Ctx, Val] => Action[Ctx, Res]) {
-  val maxLevel: Int = Integer.MAX_VALUE
-  def apply(ctx: Context[Ctx, Val], projected: Vector[ProjectedName]): Action[Ctx, Res]
-}
-
-object Projector {
-  def apply[Ctx, Val, Res](fn: (Context[Ctx, Val], Vector[ProjectedName]) => Action[Ctx, Res]) =
-    new Projector[Ctx, Val, Res] {
-      def apply(ctx: Context[Ctx, Val], projected: Vector[ProjectedName]): Action[Ctx, Res] =
-        fn(ctx, projected)
-      override def apply(ctx: Context[Ctx, Val]) = throw new IllegalStateException(
-        "Default apply should not be called on projector!")
-    }
-
-  def apply[Ctx, Val, Res](
-      levels: Int,
-      fn: (Context[Ctx, Val], Vector[ProjectedName]) => Action[Ctx, Res]) =
-    new Projector[Ctx, Val, Res] {
-      override val maxLevel: Int = levels
-      def apply(ctx: Context[Ctx, Val], projected: Vector[ProjectedName]): Action[Ctx, Res] =
-        fn(ctx, projected)
-      override def apply(ctx: Context[Ctx, Val]) = throw new IllegalStateException(
-        "Default apply should not be called on projector!")
-    }
-}
-
-case class ProjectedName(
-    name: String,
-    children: Vector[ProjectedName] = Vector.empty,
-    args: Args = Args.empty) {
-  lazy val asVector: Vector[Vector[String]] = {
-    def loop(name: ProjectedName): Vector[Vector[String]] =
-      Vector(name.name) +: (name.children.flatMap(loop).map(name.name +: _))
-
-    loop(this)
-  }
-}
 
 case class MappingDeferred[A, +B](deferred: Deferred[A], mapFn: A => (B, Vector[Throwable]))
     extends Deferred[B]
@@ -291,103 +87,6 @@ trait WithInputTypeRendering[Ctx] {
 
   def renderInputValueCompact[T](value: (_, ToInput[_, _]), tpe: InputType[T]): Option[String] =
     DefaultValueRenderer.renderInputValueCompact(value, tpe, coercionHelper)
-}
-
-case class DefaultValueParser[T](
-    schema: Schema[_, _],
-    parser: InputParser[T],
-    toInput: ToInput[T, _])
-
-object DefaultValueParser {
-  def forType[T](schema: Schema[_, _])(implicit
-      parser: InputParser[T],
-      toInput: ToInput[T, _]): DefaultValueParser[T] =
-    DefaultValueParser[T](schema, parser, toInput)
-}
-
-object DefaultValueRenderer {
-  implicit val marshaller: QueryAstResultMarshaller =
-    sangria.marshalling.queryAst.queryAstResultMarshaller
-
-  def renderInputValueCompact[T, Ctx](
-      value: (_, ToInput[_, _]),
-      tpe: InputType[T],
-      coercionHelper: ValueCoercionHelper[Ctx]): Option[String] =
-    renderInputValue(value, tpe, coercionHelper).map(marshaller.renderCompact)
-
-  def renderInputValuePretty[T, Ctx](
-      value: (_, ToInput[_, _]),
-      tpe: InputType[T],
-      coercionHelper: ValueCoercionHelper[Ctx]): Option[String] =
-    renderInputValue(value, tpe, coercionHelper).map(marshaller.renderPretty)
-
-  def renderInputValue[T, Ctx](
-      value: (_, ToInput[_, _]),
-      tpe: InputType[T],
-      coercionHelper: ValueCoercionHelper[Ctx]): Option[marshaller.Node] = {
-    val (v, toInput) = value.asInstanceOf[(Any, ToInput[Any, Any])]
-    val (inputValue, iu) = toInput.toInput(v)
-
-    if (!iu.isDefined(inputValue))
-      None
-    else
-      coercionHelper.coerceInputValue(
-        tpe,
-        Nil,
-        inputValue,
-        None,
-        None,
-        CoercedScalaResultMarshaller.default,
-        CoercedScalaResultMarshaller.default,
-        isArgument = false)(iu) match {
-        case Right(Trinary.Defined(coerced)) => Some(renderCoercedInputValue(tpe, coerced))
-        case _ => None
-      }
-  }
-
-  def renderCoercedInputValueCompact[T](value: Any, tpe: InputType[T]): String =
-    marshaller.renderCompact(renderCoercedInputValue(tpe, value))
-
-  def renderCoercedInputValuePretty[T](value: Any, tpe: InputType[T]): String =
-    marshaller.renderPretty(renderCoercedInputValue(tpe, value))
-
-  def renderCoercedInputValue(t: InputType[_], v: Any): marshaller.Node = t match {
-    case _ if v == null => marshaller.nullNode
-    case s: ScalarType[Any @unchecked] =>
-      Resolver.marshalScalarValue(
-        s.coerceOutput(v, marshaller.capabilities),
-        marshaller,
-        s.name,
-        s.scalarInfo)
-    case s: ScalarAlias[Any @unchecked, Any @unchecked] =>
-      renderCoercedInputValue(s.aliasFor, s.toScalar(v))
-    case e: EnumType[Any @unchecked] =>
-      Resolver.marshalEnumValue(e.coerceOutput(v), marshaller, e.name)
-    case io: InputObjectType[_] =>
-      val mapValue = v.asInstanceOf[Map[String, Any]]
-
-      val builder = io.fields.foldLeft(marshaller.emptyMapNode(io.fields.map(_.name))) {
-        case (acc, field) if mapValue contains field.name =>
-          marshaller.addMapNodeElem(
-            acc,
-            field.name,
-            renderCoercedInputValue(field.fieldType, mapValue(field.name)),
-            optional = false)
-        case (acc, _) => acc
-      }
-
-      marshaller.mapNode(builder)
-    case l: ListInputType[_] =>
-      val listValue = v.asInstanceOf[Seq[Any]]
-
-      marshaller.mapAndMarshal[Any](listValue, renderCoercedInputValue(l.ofType, _))
-    case o: OptionInputType[_] =>
-      v match {
-        case Some(optVal) => renderCoercedInputValue(o.ofType, optVal)
-        case None => marshaller.nullNode
-        case other => renderCoercedInputValue(o.ofType, other)
-      }
-  }
 }
 
 /** The context of a field during schema resolution.
